replace raw pointers in throdb tournament types with 'static references
These structs are self-referential, so we can't easily use owned data. The easiest thing to do is just to leak memory. This is not that great though, because it doesn't ensure that the vector of tournaments doesn't have outdated references. (I think it was leaked in the original zebra, too  though)